Fully implement basic CRUD for workSessions
- We can initialize the database connection and pass it around to
other modules that may need it.
- workSessions has been moved into its own router. Within this router,
paths are relative to where it's mounted by the app, and we've done:

  `app.use('/work_sessions', workSessions)`

- To make sure that PUT and PATCH do the same things, we can define
an `update` function that we just use for both methods.

---

* I'm doing `Object.assign` to update objects. This seems OK but it
feels a bit weird. However it seems like the other way mongoose wants
you to update documents is by always querying for it, i.e.
query-and-update. I'm already returning the object with my `id` param so
doing it this way seemed redundant.
* Next thing we could refactor out of `work_sessions.js` would be the
schema and model definitions.